+++ tests/test_console.py
@@ -28,6 +28,12 @@ def test_dumb_terminal():
     assert height == 25
 
 
+def test_soft_wrap():
+    console = Console(file=io.StringIO(), width=20, soft_wrap=True)
+    console.print("foo " * 10)
+    assert console.file.getvalue() == "foo " * 20
+
+
 @pytest.mark.skipif(sys.platform == "win32", reason="does not run on windows")
 def test_16color_terminal():
     console = Console(
@@ -451,3 +457,13 @@ def test_get_time() -> None:
     )
     assert console.get_time() == 99
     assert console.get_datetime() == datetime.datetime(1974, 7, 5)
+
+
+def test_console_style() -> None:
+    console = Console(
+        file=io.StringIO(), color_system="truecolor", force_terminal=True, style="red"
+    )
+    console.print("foo")
+    expected = "\x1b[31mfoo\x1b[0m\n"
+    result = console.file.getvalue()
+    assert result == expected